Don't rm directory if running from install dir

+++ install.sh
@@ -3,14 +3,17 @@
 # Install prerequisites if root
 if [ "$EUID" -ne 0 ]; then
   echo "Not root, skipping update and install"
+  exit
 else
   apt-get update
   apt-get -y install binutils gcc make
 fi
 
 # Delete directory if it exists and make empty directory
-rm -rf /opt/ps4sdk || true
-mkdir -p /opt/ps4sdk
+if [ "$PWD" != "/opt/ps4sdk" ]; then
+  rm -rf /opt/ps4sdk || true
+  mkdir -p /opt/ps4sdk
+fi
 
 # Build SDK
 cd libPS4 || (echo "Unable to enter subdirectory" && exit)
@@ -18,7 +21,10 @@ make
 cd ..
 
 # Copy compiled SDK
-cp -r libPS4 /opt/ps4sdk
+if [ "$PWD" != "/opt/ps4sdk" ]; then
+  cp -r libPS4 /opt/ps4sdk
+  cp install.sh /opt/ps4sdk/update.sh
+fi
 
 # Clear path from .bashrc if it's there
 sed -i "s/^\s*export PS4SDK=.*//gm" ~/.bashrc
